Add Galaxy class as replacement for xml parser of milky_way.xml
Renamed galaxy.cpp to galaxy_utils.cpp to clear filename.
